Conform to ATLAS guidelines for conference proceedings
Added CC-BY-4.0 licence as a foot note on the frontpage. Changed the conclusion
to add ATLAS and the LHC in the wording. Added the CDS link to the TDAQ TDR and
reformatted and reordered references as needed. Add reference to the LHC
in the abstract. Finally, added line numbers for review.